add code review and basic rework
